requirements.txt
fastapi
python-telegram-bot
supabase
openai
uvicorn
python-dotenv
requests
